Remove specific terms component and include it in resource info file
