+++ app/pages/resource/resource-info/ResourceInfo.js
@@ -10,7 +10,6 @@ import WrappedText from '../../../shared/wrapped-text/WrappedText';
 import { getServiceMapUrl } from '../../../utils/unitUtils';
 import ReservationInfo from '../reservation-info/ReservationInfo';
 import Equipment from '../resource-equipment/ResourceEquipment';
-import SpecificTerms from '../resource-specific-terms/ResourceSpecificTerms';
 
 function ResourceInfo({
   isLoggedIn, resource, unit, t
@@ -26,7 +25,12 @@ function ResourceInfo({
         <Panel defaultExpanded header={t('ResourceInfo.reservationTitle')}>
           <ReservationInfo isLoggedIn={isLoggedIn} resource={resource} />
         </Panel>
-        <SpecificTerms resource={resource} />
+        <Panel>
+          <h3>{t('ResourcePage.specificTerms')}</h3>
+          <Row>
+            {resource.specificTerms}
+          </Row>
+        </Panel>
         <Panel defaultExpanded header={t('ResourceInfo.additionalInfoTitle')}>
           <Row>
             <Col className="app-ResourceInfo__address" xs={6}>
